requirements.txt
absl-py==1.3.0
astunparse==1.6.3        
cachetools==5.2.0        
certifi==2022.12.7       
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
distro==1.8.0
Flask
Flask-SQLAlchemy==3.0.2  
flatbuffers==22.12.6     
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.51.1
h5py==3.7.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python
opt-einsum==3.3.0
packaging==22.0
Pillow==9.3.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-build==0.16.4
six==1.16.0
SQLAlchemy==1.4.44
termcolor==2.1.1
tensorflow
typing_extensions==4.4.0
scikit-image
matplotlib
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
customtkinter